request update: 修复图片内存泄漏问题
